+++ purchase_request/__manifest__.py
@@ -22,6 +22,7 @@
         "data/purchase_request_data.xml",
         "reports/report_purchase_request.xml",
         "views/purchase_request_view.xml",
+        "views/purchase_request_line_view.xml",
         "views/purchase_request_report.xml",
         "views/product_template.xml",
         "views/purchase_order_view.xml",
