+++ purchase_request/models/__init__.py
@@ -1,6 +1,7 @@
 from . import purchase_request_allocation
 from . import orderpoint
 from . import purchase_request
+from . import purchase_request_line
 from . import stock_rule
 from . import product_template
 from . import purchase_order
